+++ provisioners/puppet/modules/config/spec/classes/init_spec.rb
@@ -7,7 +7,7 @@ describe 'config' do
 
     system_packages = [ 'gcc', 'ruby-devel', 'zlib-devel', ]
     gems = [ 'bundler', 'io-console', ]
-    puppet_gems = [ 'nokogiri', 'ruby_aem', ]
+    puppet_gems = [ 'nokogiri', 'ruby_aem', 'ruby_aem_aws', ]
 
     system_packages.each do |pkg|
       it { is_expected.to contain_package(pkg) }