+++ test/inspec/aem_base_spec.rb
@@ -3,5 +3,9 @@ require './spec_helper'
 init_conf
 
 describe package('ruby_aem') do
-  it { should be_installed.by('gem').with_version('1.4.1') }
+  it { should be_installed.by('gem').with_version('1.4.2') }
+end
+
+describe package('ruby_aem_aws') do
+  it { should be_installed.by('gem').with_version('0.9.1') }
 end